cleanup

+++ dcpp/ActionRaw.h
@@ -32,13 +32,13 @@ struct Raw {
 	
 	Raw() : name(), raw(),id(0),time(0), enabled(false) { }
 	
-	Raw(int _id, const std::string& _name, const std::string& _raw, int _time, bool _enabled)
+	Raw(int _id, const std::string& _name, const std::string& _raw, uint64_t _time, bool _enabled)
 		: name(_name), raw(_raw),id(_id), time(_time), enabled(_enabled) { }
 	
 	GETSET(std::string, name, Name);
 	GETSET(std::string, raw, Raw);
 	GETSET(int, id, Id);
-	GETSET(int64_t, time, Time);
+	GETSET(uint64_t, time, Time);
 	GETSET(bool, enabled, Enabled);
 
 };
